+++ client/src/components/list/modal/Modal.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useSelector } from 'react-redux';
-import { RootState } from '../../store';
+import { RootState } from '../../../store';
 
 type ModalProps = {
   changeOpen: () => void;
