+++ client/src/styles/components/list/journeys/JourneyItems.styles.ts
