+++ client/src/styles/components/list/stations/StationItems.styles.ts
